refactor infinite board
